+++ changelog.md
@@ -3,6 +3,7 @@
 - Add vapor concentration and dew point to all device channels that support temperature and humidity
 - Add HmIP-FCI6 to batteries
 - Fix OperatingVoltageLevel attributes: low_bat_limit, low_bat_limit_default
+- Ignore parameters on initial load (ERROR*, RSSI*)
 
 # Version 2025.2.5 (2025-02-05)